Copy edit
Here are my suggested edits.
Notes:
On line 137, I don't understand what we mean by "paste doesn't work, so use the right mouse button to paste."
Line 149: Why don't we fix the documentation?
Line 155: You're mixing `apt` and `apt-get` commands on these next lines. I suggest using one or the other.
Line 159: Do we want to explain what Git is? I'm not sure it's necessary; if the person isn't a developer, a lot of this might be beyond them. But we provided a lot of background earlier on in the guide, so maybe.
Line 179: Seems like something we should change in the documentation or in the Singer spec.
Line 350: Do you want to define "stream"?
Line 404: What should they do next?
Line 294 and 499: Do we want to explain what the start date is for?
For the last half of the macOS section I simply copied the edited text from the PC section.